+++ packages/adp-tooling/src/index.ts
@@ -1,6 +1,9 @@
 export * from './types';
 export * from './prompts';
 export * from './preview/adp-preview';
+export * from './base/helper';
+export * from './base/validators';
+export * from './base/abap';
 export { generate, migrate } from './writer';
 export { generateChange } from './writer/editors';
 export { promptGeneratorInput, PromptDefaults } from './base/prompt';