+++ packages/create/src/validation/validation.ts
@@ -1,8 +1,8 @@
 import { getVariant, isCFEnvironment } from '@sap-ux/adp-tooling';
 import type { DescriptorVariantContent } from '@sap-ux/adp-tooling';
 import { getAppType, getWebappPath } from '@sap-ux/project-access';
-import { existsSync } from 'fs';
 import type { Editor } from 'mem-fs-editor';
+import { existsSync } from 'fs';
 import { join } from 'path';
 
 /**